Fix : 테스트 통합 및 예외사항 추가(단위/통합 테스트 처리 포함)
