+++ project/koodongha-onboarding/app/src/test/kotlin/com/example/SurveyAnswerServiceTest.kt
@@ -1,4 +1,4 @@
-package com.example.service
+package com.example
 
 import com.example.dto.SubmitAnswerRequest
 import com.example.entity.InputType
