+++ project/koodongha-onboarding/app/src/test/kotlin/com/example/SurveyApplicationTests.kt
@@ -1,0 +1,13 @@
+package com.example
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class SurveyApplicationTests {
+
+    @Test
+    fun contextLoads() {
+        
+    }
+}
