--- project/koodongha-onboarding/app/src/test/kotlin/com/example/SurveyTestApplication.kt
@@ -1,11 +1,0 @@
-package com.example
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class SurveyTestApplication
-
-fun main(args: Array<String>) {
-    runApplication<SurveyTestApplication>(*args)
-}
